Initial work on Multimodels
A multimodel takes one or more splitmodels and allows them all to work
on the same state.

Currently it is not perfect, it has the same problems of scoping as
working on global variables. Models can ruin each others state without thinking about it.

Also as they work on the state in the order that the model was put in,
that can have an impact (it changes whether a calculation was done with
v[t] or v[t-1]). The values have to be exactly the same as well, if
there was something that was a constant in one model, but changed over
time in another, it would not work out.

I think ideally state should be scoped, but linked together manually,
perhaps with an optional transformation function to manage the value.