Example of combining models
This adds a simple model that alters Pcrit based off a normal
distributed random variable. It could be a lot more complex, but this
shows the concept.

It has to set the initial value for Pcrit to be similar to the lake
model so that it could be run independently.