Switch to seperate state with links
It was not obvious how state moved between models. This separates out
the state between each model and create explicit links between the
models.

A link defines a function, that takes two states.

The link is run after every update function, to move state from the
src to the dest. It can also do an arbitrary function.